Revert "Read subscriptions to send emails from cohort table"
