+++ lambda/src/main/scala/pricemigrationengine/handlers/AmendmentHandler.scala
@@ -16,7 +16,7 @@ object AmendmentHandler extends App with RequestHandler[Unit, Unit] {
 
   val main: ZIO[Logging with AmendmentConfiguration with CohortTable with Zuora, Failure, Unit] =
     for {
-      cohortItems <- CohortTable.fetch(SalesforcePriceRiceCreationComplete, None)
+      cohortItems <- CohortTable.fetch(SalesforcePriceRiceCreationComplete)
       _ <- cohortItems.foreach(amend)
     } yield ()
