+++ lambda/src/main/scala/pricemigrationengine/handlers/NotificationEmailHandler.scala
@@ -1,38 +1,25 @@
 package pricemigrationengine.handlers
 
 import com.amazonaws.services.lambda.runtime.Context
-import pricemigrationengine.model.CohortTableFilter.AmendmentComplete
-import pricemigrationengine.model.{Failure, NotificationEmailFailure}
+import pricemigrationengine.model.Failure
 import pricemigrationengine.services._
-import zio.clock.Clock
 import zio.console.Console
-import zio.{Runtime, ZEnv, ZIO, ZLayer, clock, console}
+import zio.random.Random
+import zio.{Runtime, ZEnv, ZIO, ZLayer, console}
 
 object NotificationEmailHandler {
-  private val NotificationEmailLeadTimeDays = 30
-
-  val main: ZIO[Logging with CohortTable with SalesforceClient with Clock, Failure, Unit] = {
-    for {
-      now <- clock.currentDateTime.mapError(ex => NotificationEmailFailure(s"Failed to get time: $ex"))
-      subscriptions <- CohortTable.fetch(
-        AmendmentComplete, Some(now.toLocalDate.plusDays(NotificationEmailLeadTimeDays))
-      )
-      _ <- subscriptions.foreach { subscription =>
-        Logging.info(s"Sending notification email for subscription: ${subscription.subscriptionName}")
-      }
-    } yield ()
-  }
+  val main: ZIO[Logging with CohortTable with SalesforceClient, Failure, Unit] = ???
 
   private def env(
     loggingLayer: ZLayer[Any, Nothing, Logging]
-  ): ZLayer[Any, Any, Logging with CohortTable with SalesforceClient with Clock] = {
+  ): ZLayer[Any, Any, Logging with CohortTable with SalesforceClient] = {
     val cohortTableLayer =
       loggingLayer ++ EnvConfiguration.dynamoDbImpl >>>
         DynamoDBClient.dynamoDB ++ loggingLayer >>>
         DynamoDBZIOLive.impl ++ loggingLayer ++ EnvConfiguration.cohortTableImp ++
           EnvConfiguration.stageImp ++ EnvConfiguration.salesforceImp >>>
-        CohortTableLive.impl ++ SalesforceClientLive.impl ++ Clock.live
-    loggingLayer ++ cohortTableLayer
+        CohortTableLive.impl ++ SalesforceClientLive.impl
+    loggingLayer ++ EnvConfiguration.amendmentImpl ++ cohortTableLayer
   }
 
   private val runtime = Runtime.default
