+++ lambda/src/main/scala/pricemigrationengine/model/Failure.scala
@@ -21,5 +21,3 @@ case class SalesforceClientFailure(reason: String) extends Failure
 
 case class S3Failure(reason: String) extends Failure
 case class SubscriptionIdUploadFailure(reason: String) extends Failure
-
-case class NotificationEmailFailure(reason: String) extends Failure
