+++ lambda/src/main/scala/pricemigrationengine/services/CohortTable.scala
@@ -1,7 +1,5 @@
 package pricemigrationengine.services
 
-import java.time.LocalDate
-
 import pricemigrationengine.model._
 import zio.stream.ZStream
 import zio.{IO, ZIO}
@@ -11,8 +9,7 @@ case class CohortTableKey(subscriptionNumber: String)
 object CohortTable {
   trait Service {
     def fetch(
-        filter: CohortTableFilter,
-        beforeDateInclusive: Option[LocalDate]
+        filter: CohortTableFilter
     ): IO[CohortFetchFailure, ZStream[Any, CohortFetchFailure, CohortItem]]
 
     def put(cohortItem: CohortItem): ZIO[Any, CohortUpdateFailure, Unit]
@@ -21,10 +18,9 @@ object CohortTable {
   }
 
   def fetch(
-      filter: CohortTableFilter,
-      beforeDateInclusive: Option[LocalDate]
+      filter: CohortTableFilter
   ): ZIO[CohortTable, CohortFetchFailure, ZStream[Any, CohortFetchFailure, CohortItem]] =
-    ZIO.accessM(_.get.fetch(filter, beforeDateInclusive))
+    ZIO.accessM(_.get.fetch(filter))
 
   def put(subscription: CohortItem): ZIO[CohortTable, CohortUpdateFailure, Unit] =
     ZIO.accessM(_.get.put(subscription))
