+++ lambda/src/test/scala/pricemigrationengine/handlers/SalesforcePriceRiseCreationHandlerTest.scala
@@ -3,7 +3,6 @@ package pricemigrationengine.handlers
 import java.time.{DateTimeException, Instant, LocalDate, OffsetDateTime, ZoneOffset}
 import java.util.concurrent.TimeUnit
 
-import pricemigrationengine.StubClock
 import pricemigrationengine.model.CohortTableFilter.{EstimationComplete, SalesforcePriceRiceCreationComplete}
 import pricemigrationengine.model._
 import pricemigrationengine.services._
@@ -24,6 +23,16 @@ class SalesforcePriceRiseCreationHandlerTest extends munit.FunSuite {
     }
   )
 
+  val expectedCurrentTime = Instant.parse("2020-05-21T15:16:37Z")
+  val stubClock = ZLayer.succeed(
+    new Clock.Service {
+      override def currentTime(unit: TimeUnit): UIO[Long] = ???
+      override def currentDateTime: IO[DateTimeException, OffsetDateTime] =
+        IO.succeed(expectedCurrentTime.atOffset(ZoneOffset.of("-08:00")))
+      override def nanoTime: UIO[Long] = ???
+      override def sleep(duration: Duration): UIO[Unit] = ???
+    }
+  )
   val stubLogging = console.Console.live >>> ConsoleLogging.impl
 
   val expectedSubscriptionName = "Sub-0001"
@@ -36,9 +45,8 @@ class SalesforcePriceRiseCreationHandlerTest extends munit.FunSuite {
     ZLayer.succeed(
       new CohortTable.Service {
         override def fetch(
-          filter: CohortTableFilter,
-          beforeDateInclusive: Option[LocalDate]
-        ): IO[CohortFetchFailure, ZStream[Any, CohortFetchFailure, CohortItem]] = {
+                            filter: CohortTableFilter
+                          ): IO[CohortFetchFailure, ZStream[Any, CohortFetchFailure, CohortItem]] = {
           assertEquals(filter, EstimationComplete)
           IO.succeed(ZStream(cohortItem))
         }
@@ -106,7 +114,7 @@ class SalesforcePriceRiseCreationHandlerTest extends munit.FunSuite {
       default.unsafeRunSync(
         SalesforcePriceRiseCreationHandler.main
           .provideLayer(
-            stubLogging ++ stubConfiguration ++ stubCohortTable ++ stubSalesforceClient ++ StubClock.clock
+            stubLogging ++ stubConfiguration ++ stubCohortTable ++ stubSalesforceClient ++ stubClock
           )
       ),
       Success(())
@@ -135,7 +143,7 @@ class SalesforcePriceRiseCreationHandlerTest extends munit.FunSuite {
     )
     assertEquals(
       updatedResultsWrittenToCohortTable(0).whenSfShowEstimate,
-      Some(StubClock.expectedCurrentTime)
+      Some(expectedCurrentTime)
     )
   }
 
@@ -165,7 +173,7 @@ class SalesforcePriceRiseCreationHandlerTest extends munit.FunSuite {
       default.unsafeRunSync(
         SalesforcePriceRiseCreationHandler.main
           .provideLayer(
-            stubLogging ++ stubConfiguration ++ stubCohortTable ++ stubSalesforceClient ++ StubClock.clock
+            stubLogging ++ stubConfiguration ++ stubCohortTable ++ stubSalesforceClient ++ stubClock
           )
       ),
       Success(())
@@ -193,7 +201,7 @@ class SalesforcePriceRiseCreationHandlerTest extends munit.FunSuite {
     )
     assertEquals(
       updatedResultsWrittenToCohortTable(0).whenSfShowEstimate,
-      Some(StubClock.expectedCurrentTime)
+      Some(expectedCurrentTime)
     )
   }
 }
